[fix]QuestionBase作成,検索機能修正#163
