+++ front/src/components/Tags/Tagslist.vue
@@ -3,52 +3,50 @@
     <div class="relative inline-block text-left ml-2">
       <div class="font-bold">ジャンルから検索</div>
       <div v-for="tag in tags" :key="tag.id">
-        <div
-          class="
-            origin-top-right
-            right-0
-            mt-2
-            w-56
-            rounded-md
-            shadow-lg
-            bg-white
-            dark:bg-gray-800
-            ring-1 ring-black ring-opacity-5
-          "
-        >
-          <div class="py-1">
-            <a
-              href="#"
-              class="
-                block block
-                px-4
-                py-2
-                text-md text-gray-700
-                hover:bg-gray-100
-                hover:text-gray-900
-                dark:text-gray-100
-                dark:hover:text-white
-                dark:hover:bg-gray-600
-              "
-            >
-              <span class="flex flex-col">
-                <router-link :to="{ name: 'QuestionTag', params: { tag: tag } }">
-                  
-                <span @click="childEvent">
-                  {{ tag }}
+        <router-link :to="{ name: 'QuestionTag', params: { tag: tag } }">
+          <div
+            class="
+              origin-top-right
+              right-0
+              mt-2
+              w-56
+              rounded-md
+              shadow-lg
+              bg-white
+              dark:bg-gray-800
+              ring-1 ring-black ring-opacity-5
+            "
+          >
+            <div class="py-1">
+              <a
+                href="#"
+                class="
+                  block block
+                  px-4
+                  py-2
+                  text-md text-gray-700
+                  hover:bg-gray-100
+                  hover:text-gray-900
+                  dark:text-gray-100
+                  dark:hover:text-white
+                  dark:hover:bg-gray-600
+                "
+              >
+                <span class="flex flex-col">
+                  <span @click="childEvent">
+                    {{ tag }}
+                  </span>
                 </span>
-                </router-link>
-              </span>
-            </a>
+              </a>
+            </div>
           </div>
-        </div>
+        </router-link>
       </div>
     </div>
   </div>
 </template>
 
 <script>
-
 export default {
   data: function () {
     return {
@@ -69,10 +67,10 @@ export default {
       ],
     };
   },
- methods: {
+  methods: {
     childEvent() {
-      this.$emit('parent-event')
-    }
-  }
+      this.$emit("parent-event");
+    },
+  },
 };
 </script>
