+++ front/src/components/questions/QuestionCommentsCount.vue
@@ -1,35 +1,26 @@
 <template>
-<div>
-              <a
-              href="#"
-              class="
-                grid grid-rows-2
-                mx-auto
-                mb-3
-                py-1
-                w-4/5
-                2lg:w-3/5
-                rounded-md
-                bg-green-400
-              "
-            >
-              <div class="inline-block font-medium text-2xl text-white">
-                {{ count }}
-              </div>
-              <div
-                class="
-                  inline-block
-                  font-medium
-                  text-white
-                  mx-1
-                  text-sm
-                  lg:text-md
-                "
-              >
-                回答
-              </div>
-            </a>
-</div>
+  <div>
+    <a
+      href="#"
+      class="
+        grid grid-rows-2
+        mx-auto
+        mb-3
+        py-1
+        w-4/5
+        2lg:w-3/5
+        rounded-md
+        bg-green-400
+      "
+    >
+      <div class="inline-block font-medium text-2xl text-white">
+        {{ count }}
+      </div>
+      <div class="inline-block font-medium text-white mx-1 text-sm lg:text-md">
+        回答
+      </div>
+    </a>
+  </div>
 </template>
 
 <script>
@@ -38,15 +29,15 @@ export default {
   props: ["question"],
   data: function () {
     return {
-      count: ''
+      count: "",
     };
   },
-    created() {
+  created() {
     axios.get(`/api/v1/comments_count/${this.question}`).then((response) => {
       this.count = response.data;
       console.log(response.data);
     });
   },
-}
+};
 </script>
