+++ front/src/components/questions/QuestionStatus.vue
@@ -1,51 +1,50 @@
 <template>
-<!-- v-if="craeted" -->
-  <div class="col-span-12 lg:col-span-8" >
-                <div
-                  v-if="bestAnswer"
-                  class="
-                    inline-block
-                    rounded-full
-                    text-white
-                    bg-red-400
-                    duration-300
-                    text-xs
-                    font-bold
-                    mr-1
-                    md:mr-2
-                    mb-2
-                    px-2
-                    md:px-4
-                    py-1
-                    opacity-90
-                  "
-                >
-                  解決済
-                </div>
-                <div
-                  v-else
-                  href="#"
-                  class="
-                    inline-block
-                    rounded-full
-                    text-white
-                    bg-blue-400
-                    duration-300
-                    text-xs
-                    font-bold
-                    mr-1
-                    md:mr-2
-                    mb-2
-                    px-2
-                    md:px-4
-                    py-1
-                    opacity-90
-                  "
-                >
-                  回答受付中
-                </div>
-              </div>
-            
+  <!-- v-if="craeted" -->
+  <div class="col-span-12 lg:col-span-8">
+    <div
+      v-if="bestAnswer"
+      class="
+        inline-block
+        rounded-full
+        text-white
+        bg-red-400
+        duration-300
+        text-xs
+        font-bold
+        mr-1
+        md:mr-2
+        mb-2
+        px-2
+        md:px-4
+        py-1
+        opacity-90
+      "
+    >
+      解決済
+    </div>
+    <div
+      v-else
+      href="#"
+      class="
+        inline-block
+        rounded-full
+        text-white
+        bg-blue-400
+        duration-300
+        text-xs
+        font-bold
+        mr-1
+        md:mr-2
+        mb-2
+        px-2
+        md:px-4
+        py-1
+        opacity-90
+      "
+    >
+      回答受付中
+    </div>
+  </div>
 </template>
 
 <script>
@@ -54,18 +53,19 @@ export default {
   props: ["question"],
   data: function () {
     return {
-      bestAnswer: '',
-      
+      bestAnswer: "",
     };
   },
   created() {
-  axios.get(`/api/v1/bestanswer/${this.question.id}`).then((response) => {
-      this.bestAnswer = response.data;
-      console.log(response.data);
-    }).catch((error) => {
-          console.log(error);
-        });
-    
-  }
-}
+    axios
+      .get(`/api/v1/bestanswer/${this.question.id}`)
+      .then((response) => {
+        this.bestAnswer = response.data;
+        console.log(response.data);
+      })
+      .catch((error) => {
+        console.log(error);
+      });
+  },
+};
 </script>
